+++ agent_workbench/package/__init__.py
@@ -10,10 +10,14 @@ from __future__ import annotations
 from agent_workbench.package.exceptions import PackageValidationError
 from agent_workbench.package.loader import PackageLoader
 from agent_workbench.package.manifest import PackageManifest
+from agent_workbench.package.package_info import PackageInfo
+from agent_workbench.package.registry import PackageRegistry
 
 
 __all__ = [
+    "PackageInfo",
     "PackageLoader",
     "PackageManifest",
+    "PackageRegistry",
     "PackageValidationError",
 ]